+++ oysterpack-uid/CHANGELOG.md
@@ -6,10 +6,18 @@ This project adheres to [Semantic Versioning](http://semver.org/), as described 
 
 ## \[Unreleased\]
 
-## \[0.2.3\] 2018-12-24
+## \[0.2.3\] 2018-12-27
 - re-organized project structure
-- added benchmark tests
-- changed ULID serde serialization to u128 instead of string for performance boost
+
+### Added
+- benchmark tests
+- re-export ulid and domain attributes from oysterpack_uid_macros
+
+### Changed
+- ULID serde serialization to u128 instead of string for performance boost
+
+### Removed
+- TypedULID 
 
 ## \[0.2.2\] 2018-12-07
 - upgraded to rust 2018 edition